pyqt_hello/requirements.txt
click==8.1.8
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.10.0
python-dotenv==1.1.0
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
pillow==11.2.1
